Removing , default to creation, and remove stream handler layer
