+++ spring-cloud-stream-binder-kinesis-core/src/main/java/org/springframework/cloud/stream/binder/kinesis/properties/KinesisBinderConfigurationProperties.java
@@ -26,13 +26,4 @@ import org.springframework.boot.context.properties.ConfigurationProperties;
 @ConfigurationProperties(prefix = "spring.cloud.stream.kinesis.binder")
 public class KinesisBinderConfigurationProperties {
 
-	private boolean autoCreateStreams = true;
-
-	public boolean getAutoCreateStreams() {
-		return autoCreateStreams;
-	}
-
-	public void setAutoCreateStreams(boolean autoCreateStreams) {
-		this.autoCreateStreams = autoCreateStreams;
-	}
 }
