--- spring-cloud-stream-binder-kinesis-core/src/main/java/org/springframework/cloud/stream/binder/kinesis/provisioning/KinesisStream.java
@@ -1,26 +1,0 @@
-package org.springframework.cloud.stream.binder.kinesis.provisioning;
-
-/**
- * Represents a Kinesis stream.
- *
- * @author Jacob Severson
- */
-public class KinesisStream {
-
-	private final String name;
-
-	private final Integer shards;
-
-	public KinesisStream(String name, Integer shards) {
-		this.name = name;
-		this.shards = shards;
-	}
-
-	public String getName() {
-		return name;
-	}
-
-	public Integer getShards() {
-		return shards;
-	}
-}
