+++ spring-cloud-stream-binder-kinesis/src/test/java/org/springframework/cloud/stream/binder/kinesis/KinesisTestBinder.java
@@ -22,7 +22,6 @@ import org.springframework.cloud.stream.binder.ExtendedProducerProperties;
 import org.springframework.cloud.stream.binder.kinesis.properties.KinesisBinderConfigurationProperties;
 import org.springframework.cloud.stream.binder.kinesis.properties.KinesisConsumerProperties;
 import org.springframework.cloud.stream.binder.kinesis.properties.KinesisProducerProperties;
-import org.springframework.cloud.stream.binder.kinesis.provisioning.KinesisStreamHandler;
 import org.springframework.cloud.stream.binder.kinesis.provisioning.KinesisStreamProvisioner;
 import org.springframework.context.support.GenericApplicationContext;
 import org.springframework.integration.codec.kryo.PojoCodec;
@@ -38,10 +37,9 @@ public class KinesisTestBinder
 
 	public KinesisTestBinder(AmazonKinesisAsync amazonKinesis,
 			KinesisBinderConfigurationProperties kinesisBinderConfigurationProperties) {
-		KinesisStreamHandler kinesisStreamHandler = new KinesisStreamHandler(amazonKinesis);
 
 		KinesisStreamProvisioner provisioningProvider =
-				new KinesisStreamProvisioner(kinesisStreamHandler, kinesisBinderConfigurationProperties);
+				new KinesisStreamProvisioner(amazonKinesis, kinesisBinderConfigurationProperties);
 
 		KinesisMessageChannelBinder binder =
 				new KinesisMessageChannelBinder(amazonKinesis, kinesisBinderConfigurationProperties, provisioningProvider);
